feat(postbuildstepper): query source branch from github API; improve test
