+++ applications/postbuildstepper/Cargo.toml
@@ -13,3 +13,5 @@ log = "0.4.22"
 pcre2 = "0.2.9"
 serde_json = "1.0.132"
 tempfile = "3.13.0"
+reqwest = { version = "0.12", features = ["json"] }
+tokio = { version = "1", features = ["full"] }
