+++ secrets/buildbot-nix-0/secrets.yaml
@@ -19,7 +19,7 @@ buildbot-github-webhook-secret: ENC[AES256_GCM,data:uGBf61+hqBwRzbXhdzaJqR23smFB
 #
 buildbot-nix-workers: ENC[AES256_GCM,data:PfYOhXRZDjBl5pmZjTXr1a4fM5Fxqcnb3t9kdk37dmGMx7dqho7ke2546PiRJQSoJBZAe6jjxTgxI5UmU1VErItc0oJT3xSNDpuD8QqVuSQTZ+guAU4idRR6Hg==,iv:VA1NAGmsXC12XomWuIy1C9POPYFktJC9g5KsETFVVvM=,tag:XsUCniGlUg7IlJtSUkUmWw==,type:str]
 buildbot-nix-worker-password: ENC[AES256_GCM,data:96s+7HzGRAznb/3757jS+JnOXBI2yapm5FNgyferFs4=,iv:7cnQMrmGD0bhz9NWnM5n8G0WbhPfYo8g0usdd2IFRiA=,tag:FPHTk0Om17L1C4GzcK2hjQ==,type:str]
-holo-host-github-environment-secrets: ENC[AES256_GCM,data:q+CD/hSJEYFYoUcoR/6uCUvRRUNwIEAnMlNknX6KhXwhMm043cDfscQ0W5cKWy1SPxq9HLY7YsDXTMwSTT1wpEL+Ai9PfYKUkC81mLJUrf8RMN7ydYUkKMO5TzAk/c8izmHyocqbWPsYHkLUK8Nr,iv:nA7PTg3Nq/sF2nanzPg7+dqpcEZmSrT02f7L1qdd66Q=,tag:RWSu/GgCO0pOY4LQOj/UQQ==,type:str]
+holo-host-github-environment-secrets: ENC[AES256_GCM,data:TqTEy5aTNU9ew0a9TG6JzWIjJX+2nzFvKht0449G5aELbsM+K/TLZQIkjH1TAyH/UZRkkAqD/ipKs8KQl0Ss3sJCgfF4tGxJaFbMxdK+JXTQmYmPmxZKQhseb3AebDrGlxj3GvY6hfKCWcYJhAApyecr4nR9tA64UVY/NnwAvpuyniOGdGQFq1Rte/1VggB9ebfYNjnwSob+dKoOoGTlwGB3uGnsAAXpCG/xoR8Jc3QwDEnl/QsmsgGN27Q6mFUYdV6NFEzkVejvTdHeHUS9PZu9k4mYRxe65SRYLFDvJ0E2N9cwFG0=,iv:VdHO6aNPKknbqKM81hWET/LXGn5G2aH5WiReka3qOkk=,tag:2/+xOItNKHHesK89btMcnQ==,type:str]
 holo-host-aws-shared-credentials: ENC[AES256_GCM,data:RCoLA7oYGersXqIZtehqCdWCFI9D2drY5OqSg9SCqkeFP7xbTw1t2pge6vFktZ0PWr84n96h52jY4t7M4hWNYiEb6czrL7zF8cacOlL335HRiDLacv4IS2Lk00AEvEP9mmE2OhHXJ/2c6ZN17OuVR+Pp4nSscStwciM866U9t5+OKehzzLw=,iv:/E8Qeo5VD20JteihvZgR057rCArudSUt2+gdAx4Eq4c=,tag:7MF9eNozQjvfNdEzu4vL0Q==,type:str]
 sops:
     kms: []
@@ -45,8 +45,8 @@ sops:
             RU8yUUhwcmFpTmlRQzVHcmJRUVhNZWcKDNgiTWktwmVLKvrYhm7OUNgTQa1y+Bdc
             N4N+lnkn657Ji21zr1X9kDsvJoXASZXj2qkwuHzIpJgz3SMri0F2Ow==
             -----END AGE ENCRYPTED FILE-----
-    lastmodified: "2024-10-19T18:04:25Z"
-    mac: ENC[AES256_GCM,data:AaGps5HfEOgOruIwm1UbaMUdusA0Y8SpIAhaAuRtfbCOOisQGUQYuWkKIw9Fqm5A97lJcqrKEkYg1rI6aRXUYg070gNYNT13YgeAnD2ZqZZnHY1izm2iO0bOJK6tBA4pHYdcpe+NnSYkAF7FgbiJkpTHHBD9luI1eHkWM+x9pXg=,iv:u8fZOJLR7LnkxHAw6mYmY1pMEXkP5+rzjG+uMvkhSik=,tag:5aduXpzBCiRYpwodgCXGcg==,type:str]
+    lastmodified: "2024-11-05T13:56:03Z"
+    mac: ENC[AES256_GCM,data:mQDqAEbD7K/77lbuGeVsnxs37nErWzucZcls6E4O6QcvdJAtKRqYj6bmJmhQ7l/s6PU577I+KH9PoySVEh+l7KGo1pwFLP/BLiinTArKFzxVHZee1hNpYDVMPz92+w4sNUZOarUMkjT9cdCYEDuneF5uQcRSaYcO2RaYDkK4dCU=,iv:7f+BPaqq8R74mOGvRwqZO0mF2tC6iTFWWRD9et7GK8Q=,tag:7aemFQNcT8UudGhrO8g8Rw==,type:str]
     pgp:
         - created_at: "2024-10-07T16:27:41Z"
           enc: |-
